building full example
