+++ example/lib/graphql/hero_for_episode.gql
