+++ src/helpers/expected-generated-file-for.js
@@ -1,0 +1,7 @@
+import { basename, extname } from "path";
+
+export default function expectedGeneratedFileFor(operations) {
+  let fileName = operations[0].originalFile;
+  const base = basename(fileName, extname(fileName));
+  return `${base}.g.dart`;
+}
