+++ src/templates/documents/documents.handlebars
@@ -3,6 +3,8 @@
 import 'package:meta/meta.dart';
 import 'package:json_annotation/json_annotation.dart';
 
+part '{{ expectedGeneratedFileFor operations }}';
+
 abstract class ToJson {
   Map<String, dynamic> toJson();
 }
